+++ src/Microsoft.AspNetCore.Http.Connections.Client/Internal/DefaultTransportFactory.cs
@@ -5,7 +5,6 @@ using System;
 using System.Net.Http;
 using Microsoft.AspNetCore.Http.Connections.Client;
 using Microsoft.AspNetCore.Http.Connections.Client.Internal;
-using Microsoft.AspNetCore.Http.Connections.Internal;
 using Microsoft.Extensions.Logging;
 
 namespace Microsoft.AspNetCore.Http.Connections.Client.Internal
