+++ src/Microsoft.AspNetCore.Http.Connections.Client/Internal/Utils.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.AspNetCore.Http.Connections.Client
+namespace Microsoft.AspNetCore.Http.Connections.Client.Internal
 {
     internal static class Utils
     {
