+++ src/Microsoft.AspNetCore.Http.Connections.Common/AvailableTransport.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 
-namespace Microsoft.AspNetCore.Http.Connections.Internal
+namespace Microsoft.AspNetCore.Http.Connections
 {
     public class AvailableTransport
     {
