+++ src/Microsoft.AspNetCore.Http.Connections.Common/NegotiationResponse.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 
-namespace Microsoft.AspNetCore.Http.Connections.Internal
+namespace Microsoft.AspNetCore.Http.Connections
 {
     public class NegotiationResponse
     {
