+++ src/Microsoft.AspNetCore.SignalR.Client.Core/Internal/InvocationRequest.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.SignalR.Protocol;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.AspNetCore.SignalR.Client
+namespace Microsoft.AspNetCore.SignalR.Client.Internal
 {
     internal abstract class InvocationRequest : IDisposable
     {
