+++ src/Microsoft.AspNetCore.SignalR.Core/HubConnectionStore.cs
@@ -6,7 +6,7 @@ using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 
-namespace Microsoft.AspNetCore.SignalR.Internal
+namespace Microsoft.AspNetCore.SignalR
 {
     public class HubConnectionStore
     {
