+++ src/Microsoft.AspNetCore.SignalR.Core/Hub`T.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Microsoft.AspNetCore.SignalR.Internal;
+
 namespace Microsoft.AspNetCore.SignalR
 {
     public class Hub<T> : Hub where T : class
