+++ src/Microsoft.AspNetCore.SignalR.Core/Internal/SignalRMarkerService.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Extensions.DependencyInjection
+namespace Microsoft.AspNetCore.SignalR.Internal
 {
     internal class SignalRMarkerService
     {
