+++ src/Microsoft.AspNetCore.SignalR.Core/Internal/TypedHubClients.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 
-namespace Microsoft.AspNetCore.SignalR
+namespace Microsoft.AspNetCore.SignalR.Internal
 {
     internal class TypedHubClients<T> : IHubCallerClients<T>
     {
