+++ src/Microsoft.AspNetCore.SignalR.Core/SerializedHubMessage.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.AspNetCore.SignalR.Protocol;
 
-namespace Microsoft.AspNetCore.SignalR.Internal
+namespace Microsoft.AspNetCore.SignalR
 {
     /// <summary>
     /// This class is designed to support the framework. The API is subject to breaking changes.
