+++ test/Microsoft.AspNetCore.SignalR.Client.Tests/ResponseUtils.cs
@@ -7,7 +7,6 @@ using System.Net;
 using System.Net.Http;
 using Microsoft.AspNetCore.Connections;
 using Microsoft.AspNetCore.Http.Connections;
-using Microsoft.AspNetCore.Http.Connections.Internal;
 using Newtonsoft.Json;
 
 namespace Microsoft.AspNetCore.SignalR.Client.Tests
